Fixed the tests.
